requirements.txt
requests
flask
py-healthcheck
pytest
prometheus-flask-exporter